phase3/backend/requirements.txt
SQLModel
fastapi
alembic
psycopg[binary]
apscheduler
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai
aiomcp